requirements.txt
bcrypt==3.2.2
certifi
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
et-xmlfile==1.1.0
dnspython==2.2.1
Flask==2.1.3
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.21.6
openpyxl==3.0.10
pandas==1.3.5
pycparser==2.21
pymongo==4.2.0
pymongo[srv]
python-dateutil==2.8.2
password-strength==0.0.3.post2
pytz==2022.1
regex==2022.7.25
requests==2.28.1
six==1.16.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
vaderSentiment==3.3.2
Werkzeug==2.2.0
zipp==3.8.1
